+++ jsolex/src/main/resources/whats-new.md
@@ -1,6 +1,7 @@
 # Welcome to JSol'Ex {{version}}!
 
-- [What's New in Version 4.2.1](#whats-new-in-version-4-2-1) - Bug fixes
+- [What's New in Version 4.2.2](#whats-new-in-version-4-2-2) - Faster script execution
+- [Version 4.2.1](#whats-new-in-version-4-2-1) - Bug fixes
 - [Version 4.2.0](#whats-new-in-version-4-2-0) - Script repositories, Java 25, GIF support, file format settings
 - [Version 4.1.4](#whats-new-in-version-4-1-4) - Bug fixes and improvements
 - [Version 4.1.3](#whats-new-in-version-4-1-3) - Improved banding correction
@@ -9,6 +10,10 @@
 - [Version 4.1.0](#whats-new-in-version-4-1-0) - User-defined presets, collage creation
 - [Version 4.0.1](#whats-new-in-version-4-0-1) - Language selection, bug fixes
 - [Version 4.0.0](#whats-new-in-version-4-0-0) - Enhanced UI, BASS2000 integration
+
+## What's New in Version 4.2.2
+
+- Faster script execution by parallelizing independent expressions
 
 ## What's New in Version 4.2.1